+++ app/utils.js
@@ -382,6 +382,15 @@ function getMinerFromCoinbaseTx(tx) {
 					}
 				}
 			}
+
+			for (var blockHash in miningPoolsConfig.block_hashes) {
+				if (blockHash == tx.blockhash) {
+					var minerInfo = miningPoolsConfig.block_hashes[blockHash];
+					minerInfo.identifiedBy = "known block hash '" + blockHash + "'";
+
+					return minerInfo;
+				}
+			}
 		}
 	}