+++ CHANGELOG.md
@@ -28,6 +28,7 @@
     * Min / Max tx sizes
 * New tool `/block-stats` for viewing summarized block data from recent blocks
 * New tool `/mining-summary` for viewing summarized mining data from recent blocks
+* Change `/mempool-summary` to load data via ajax (UX improvement to give feedback while loading large data sets)
 * Zero-indexing for tx inputs/outputs (#173)
 * Labels for transaction output types
 * Configurable UI "sub-header" links